add submodules

+++ submodules/len2/permute_mul/convnext_base_13.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale13 = torch.rand(torch.Size([512, 1, 1]))
+        self.layer_scale13 = torch.rand(torch.Size([512, 1, 1])).to(torch.float32)
 
     def forward(self, x168):
         x169=torch.permute(x168, [0, 3, 1, 2])
