+++ submodules/len2/permute_mul/convnext_base_18.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale18 = torch.rand(torch.Size([512, 1, 1]))
+        self.layer_scale18 = torch.rand(torch.Size([512, 1, 1])).to(torch.float32)
 
     def forward(self, x223):
         x224=torch.permute(x223, [0, 3, 1, 2])
