+++ submodules/len2/permute_mul/convnext_base_6.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale6 = torch.rand(torch.Size([512, 1, 1]))
+        self.layer_scale6 = torch.rand(torch.Size([512, 1, 1])).to(torch.float32)
 
     def forward(self, x91):
         x92=torch.permute(x91, [0, 3, 1, 2])
