+++ submodules/len2/permute_mul/convnext_large_10.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale10 = torch.rand(torch.Size([768, 1, 1]))
+        self.layer_scale10 = torch.rand(torch.Size([768, 1, 1])).to(torch.float32)
 
     def forward(self, x135):
         x136=torch.permute(x135, [0, 3, 1, 2])
