+++ submodules/len2/permute_mul/convnext_large_15.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale15 = torch.rand(torch.Size([768, 1, 1]))
+        self.layer_scale15 = torch.rand(torch.Size([768, 1, 1])).to(torch.float32)
 
     def forward(self, x190):
         x191=torch.permute(x190, [0, 3, 1, 2])
