+++ submodules/len2/permute_mul/convnext_large_9.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale9 = torch.rand(torch.Size([768, 1, 1]))
+        self.layer_scale9 = torch.rand(torch.Size([768, 1, 1])).to(torch.float32)
 
     def forward(self, x124):
         x125=torch.permute(x124, [0, 3, 1, 2])
