+++ submodules/len2/permute_mul/convnext_small_25.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale25 = torch.rand(torch.Size([384, 1, 1]))
+        self.layer_scale25 = torch.rand(torch.Size([384, 1, 1])).to(torch.float32)
 
     def forward(self, x300):
         x301=torch.permute(x300, [0, 3, 1, 2])
