+++ submodules/len2/permute_mul/convnext_small_33.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale33 = torch.rand(torch.Size([768, 1, 1]))
+        self.layer_scale33 = torch.rand(torch.Size([768, 1, 1])).to(torch.float32)
 
     def forward(self, x394):
         x395=torch.permute(x394, [0, 3, 1, 2])
