+++ submodules/len2/permute_mul/convnext_tiny_0.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale0 = torch.rand(torch.Size([96, 1, 1]))
+        self.layer_scale0 = torch.rand(torch.Size([96, 1, 1])).to(torch.float32)
 
     def forward(self, x13):
         x14=torch.permute(x13, [0, 3, 1, 2])
