+++ submodules/len2/permute_mul/convnext_tiny_3.py
@@ -12,7 +12,7 @@ import operator
 class M(torch.nn.Module):
     def __init__(self):
         super(M, self).__init__()
-        self.layer_scale3 = torch.rand(torch.Size([192, 1, 1]))
+        self.layer_scale3 = torch.rand(torch.Size([192, 1, 1])).to(torch.float32)
 
     def forward(self, x52):
         x53=torch.permute(x52, [0, 3, 1, 2])
